Gathering feedback about flat mode icons
No worries, this is not supposed to stay as default. I just enabled falt
icons in the default theme in order to get feedback about technical
issues as-well as comments about the look-and-feel. Send feedback to the
patch author's email.

The plan so far is to revert this mid-February, and just have flat mode
icons enabled for some alternative themes.

Change-Id: I371f4ede432597ad7f1bc4b9ae661c9f2bc8df3e
Reviewed-by: Alessandro Portale <alessandro.portale@theqtcompany.com>