runAsync: Fix issue with deleting thread
deleteLater schedules the event in the current thread, so we
may not use it directly on a thread from within the thread itself...
Instead connect the threads finished signal to its deleteLater slot.
To avoid problems if runAsync is not run from the main thread,
move the thread to the main thread before starting it.

Task-number: QTCREATORBUG-15688
Change-Id: Iea003d00c58d1f921fb0b4ddf9cf67dcb1379833
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>