Core: Change shortcut for maximizing outputpane...
...and limit the shortcuts for the outputpanes.

An output pane gets it shortcut by the ordering number it's added
to Qt Creator. Depending on the number of enabled plugins we now
have more output panes than available digits.
Alt+9 (or Cmd+9 on OSX) had been used already for maximizing the
outputpane, so changing this shortcut and ensure to use only the
available digits for the outputpanes.

Task-number: QTCREATORBUG-15356
Change-Id: I41a9e3df11a1abde3ec4c5d4a13e932bec024154
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>