mapReduce: Add remaining API use cases from QtConcurrent
- simpler mapReduce
- blocking mappedReduced
- map with iterators
- blocking mapped

The special QtConcurrent::map for modifying the container is not needed
(it simply is Utils::map with a map function that takes a reference to a
value, and either non-const iterators or a std::ref container).

The special (blocking)filter((ed)Reduced) is simply done with the
Utils::map(ped)/map(ped)Reduce(d) functions with a map function that
takes a QFutureInterface and only reports values if the value should not
be filtered out. That way it can also be used as a
"filteredMappedReduced".

Change-Id: Id8dcbb7277be4d0e27b9e1691df35535dd481924
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>