mapReduce: Fix MSVC2013 build
Looks like it is a problem that the void template specialization
for dummyReduce didn't have operator(), but the non-specialized one
had, so resultType<dummyReduce> failed.
It also seems that passing function pointers (to template functions?)
morph into function references somewhere, so we need to sprinkle
std::decays over the place, which is the right thing anyhow, since
they are decay-copied into the thread eventually.

Change-Id: I832474d1a6c3ecef55bfac9339d5e18b6ce910cc
Reviewed-by: Alessandro Portale <alessandro.portale@theqtcompany.com>
Reviewed-by: Christian Stenger <christian.stenger@theqtcompany.com>