Application wizard for Qt.labs.control based applications
Let's make it easier to create applications with Qt labs controls. This
creates a tabbed navigation style. We can add options to the wizard to
create drill-down and other common navigation styles. Also an option
for the controls style (material, universal, etc.). And perhaps a
checkbox to include the virtual keyboard.

Change-Id: I9672ab12aa616e7ce08e90eb244fad525e5ab2b2
Reviewed-by: Alessandro Portale <alessandro.portale@theqtcompany.com>