QbsProject: Use filesGeneratedFrom() to find files for ExtraCompilers
qbs::Project::generatedFiles() without tags retrieves all generated
files recursively. That is not what ExtraCompiler expects.

This way it should do the same as the other project types. It should
only retrieve the first level of generated files. However, as
filesGeneratedFrom() is currently unimplemented in QbsProject, in fact
it doesn't retrieve anything and the feature is disabled for now.

Change-Id: I8ba0b5e9310a55fcd348b3547d8ff87d97a1b498
Task-number: QTCREATORBUG-15796
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>