Add the "Dark frame" Qt Creator theme
http://blog.qt.io/blog/2015/02/09/qt-creator-flat-style-ui-design-new-light-theme/
defines a dark and a light theme. We start by implementing the dark one.

"Dark frame" stands for dark toolbar areas and light content areas.
The name is silly, and we need to find a proper one before the release
(we will most likely keep the silly name, though).

Since the existing "dark" theme is much darker and has a dark content
area, we don't want to simply replace that.

Change-Id: If2197c4a07ae1e374be4449b6ed0383cdbe314b3
Reviewed-by: Alessandro Portale <alessandro.portale@theqtcompany.com>