Themes: Add a theme flag for "default" or "flat" ProjectsMode
I'd like a "flat" style to have a "default" projects mode.
This patch introduces a FlatProjectsMode. The painting code of the
widgets in the project mode follow that instead of the
Theme::WidgetStyle.

In the course of it, DoubleTabWidget does not imitate the dummy toolbar
anymore, but calls the style painting code, instead. That makes the
theme color DoubleTabWidget1stEmptyAreaBackgroundColor obsolete, so that
key disappears from the Theme enum and also from the creatorthemes.

Change-Id: Ia1479c761f61753d6738a43bbde368bf0b8814b2
Reviewed-by: Tobias Hunger <tobias.hunger@theqtcompany.com>