ManhattanStyle: Fix "QFSFileEngine::open: No file name specified"
It apparently was a bad idea to let QImage load an empty string in case
of flat style. In order to fix that, we simplyfy the code by making the
images static locals in the scope for non-flat painting.

That has the additional advantage that they are only loaded once.

Change-Id: I8acb2ff09fe77cc708be2670cddaa41e23c9afa7
Reviewed-by: Orgad Shaneh <orgads@gmail.com>