TextEditor: Fix tab indentation if single line is selected
If single line was indented with tab, and the insertion point was at the
beginning of the line, the selection moved away from the beginning of
the line. Pressing tab a second time then deletes the line, because the
line is no longer completely selected.

This behavior was also not consistent to the case where the insertion
point is not at the beginning of the line.
The patch makes sure that the selection stays at the beginning of the
line.

Task-number: QTCREATORBUG-8421
Change-Id: I9847c2fe41e2e4f993f5e486c071acc112adb45d
Reviewed-by: David Schulz <david.schulz@theqtcompany.com>