Help: Set sensible fallback fonts
The default fallback font was arbitrary. So far this was not a big issue
because it was basically never used.

Set sensible fallback fonts on the different platforms.
Also make sure that we can see in the settings if the default fallback
font has been changed by the user, and also separate the setting for the
different font properties. Since we did not do this before, we now have
to reset the setting for all existing users, so they benefit from the
more sensible default.

Task-number: QTCREATORBUG-15887
Change-Id: I16419f54c300580d5c9a9b19722aacf790ef66fc
GPush-Base: 44820dae1326476a6e1aa693421c61ae67bb13b0
Reviewed-by: Kai Koehne <kai.koehne@theqtcompany.com>