Remove code duplication for copying resources.
qtcreatordata.pri was created for "conditional" copying of resources to the
build tree. Adapt it a bit and use it for the "unconditionally" copied
resources as well.
A side effect is, that the unconditionally copied resources now are
also installed file by file instead of the directory as a whole,
which doesn't make a difference in the end result though.

Change-Id: I6da3eeaadcb48c19987858bbb8c2d42ee149f6f3
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@theqtcompany.com>
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>