Redone debug start/stop/interrupt/continue icons
Currently, these icons have a big "bug beetle" and a small action icon
as overlay. According to user feedback, the relevant element (the
action) has too little visibility.

This change turns the proportions around, so that the action icon is big
and the "bug beetle" is a small overlay.

Also the analyzer run icon is handled in the same way.

Change-Id: I257c66ec5892c8aabaf8a9884795ae410e793017
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>