Clang Static Analyzer: Workaround analyzing MSVC2015 projects with clang 3.8.0
Clang 3.8.0 does not detect the MSVC2015 build environment properly. Running

  $ clang.exe -v -fsyntax-only empty.cpp

shows

  -fms-compatibility-version=18 (=MSVC2013)

instead of the expected

  -fms-compatibility-version=19 (=MSVC2015)

Workaround this by setting the version explicitly for MSVC2015 toolchains. This
fixes parse issues with char16_t and friends.

We do not explicitly check for the clang version since clang 3.6.2 is broken
for MSVC2015 anyway and we do not know whether clang 3.8.1 will come with a
fix.

Task-number: QTCREATORBUG-15940
Change-Id: Ibe61af25f6d74f134cc9f9681db5a4d19ac67417
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>