ProjectExplorer: Handle imperfect build systems
Handle buildsystems with imperfect knowledge about what they
build more gracefully.

Leave RunConfigurations around if the build system can not tell
whether or not the relevant target was configured out of the
current build or completely removed from the build system.

This patch should not change the observable behavior of the code.

Change-Id: I0ddb4bd3966d184a4b5c9818e68adb107047ed1f
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>
Reviewed-by: Tim Jenssen <tim.jenssen@theqtcompany.com>