Debugger: Do setParent(0) on disabled dock widgets
When disabling a dock widget, we don't want it to show up in the Views
menu anymore and we don't want it to interfere with other views.
setParent(0) accomplishes this. However, when showing a different
perspective we need to properly parent all the widgets that belong to
it.

Task-number: QTCREATORBUG-16027
Change-Id: I68e8cd4e6ade1dd8b23fb789f03edc39671aee91
Reviewed-by: hjk <hjk@theqtcompany.com>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>