Help browser: Fix overlapping images in old docs
This works around an issue in Qt where line-height will be
interpreted as an absolute height, rather than the minimum,
so if images are larger than the specified size, they will
overlap the text.

The work around is just to change all instances of FixedHeight
to MinimumHeight until the default can be changed in Qt.

Task-number: QTBUG-51962
Change-Id: I343d30c539d434301866a7a659ef5fc300c364d6
Reviewed-by: Eike Ziller <eike.ziller@theqtcompany.com>