RemoteLinux: Relax port gathering
We might encounter the situation that protocol is given as IPv6
but the consumer of the free port information decides to open
an IPv4(only) port. As a result the next IPv6 scan will
report the port again as open (in IPv6 namespace), while the
same port in IPv4 namespace might still be blocked, and
re-use of this port fails.

Err on the safe side, and consider ports taken in either space
as blocked.

Change-Id: I2e4be40ab4df5398e26e197c12408efe905b1a2f
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>