Xcodebuild parser: Improve performance
QRegExp is extremely ineffecient when capturing everything at the start,
but not matching the end of the expression (e.g. 200+ms when parsing for
signature change message in the relatively long xcodebuild compile
output line).
QRegularExpression is much faster with the same expression (40ms with
the same example), but just checking for "endsWith" and getting the
first part of the line manually is the fastest.

Change-Id: Id5f0378e38f0edeba24036c6e9519c8107e4a361
Task-number: QTCREATORBUG-15613
Reviewed-by: Vikas Pachdha <vikas.pachdha@theqtcompany.com>
Reviewed-by: Tobias Hunger <tobias.hunger@qt.io>