QmlProfiler: Always enable the "debug messages" feature
We always have at least one debug message: The one the debug server
itself generates. So, let's unconditionally mark the feature as
recorded when loading data. As the debug messages arrive in a different
debug client, the trace client doesn't know about them otherwise,
and if the feature is not marked as recorded you cannot hide the
respective category with the filter menu.

Change-Id: I68925697486a9461d66560b10947fde70b8a6730
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>