Timeline: Aggregate events when there are too many
With this change, if you have more than 1M events in a category, the
ones with the shortest distances from one another will be drawn
together, in grey instead of their respective colors. You can zoom in
or hover-select them to see what they are.

This greatly reduces memory consumption in those cases, but it requires
some more book keeping in order to identify the events to be
aggregated.

Task-number: QTCREATORBUG-14983
Change-Id: If54d6d7efe5f720aa72003b2c307ea9562c911a1
Reviewed-by: Christian Kandeler <christian.kandeler@theqtcompany.com>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>