Beautifier: Fix clang-format file style regression
Commit 3310fbc8f2f00aacf4f3c3bb373f49b1549db8bd introduced regression replacing
-assume-filename=%file with -assume-filename=%filename, that caused only file
name part to be passed to clang-format, not whole path to the buffer.
Effectively when formatting the buffer with style set to "File", clang-format
was unable to determine related style file placement (eg. within the project
tree) anymore.

This reverts back to use %file, not %filename, making clang-format "File" style
work again as expected.

Change-Id: I5e7ae3d1186b1e7c33dc89a7d60aacc0f0671e51
Reviewed-by: Lorenz Haas <lorenz.haas@histomatics.de>
Reviewed-by: Orgad Shaneh <orgads@gmail.com>