Clang: Fix accepting absolute path clang executable
Commit 506fc40a319bc785cb6fadfb0598aa9a3d0213ae broke running the analyzer with
absolute paths.

If the clang executable name is provided, the ".exe" suffix is not needed since
we get the absolute path with Utils::Environment::searchInPath(). However, for
absolute path we need to ensure the presence of the suffix in order to point to
an actually existing file.

Also, compare case insensitive.

Task-number: QTCREATORBUG-16234
Change-Id: I9f39386c099cc544297239eca56a95c2e7430555
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Alessandro Portale <alessandro.portale@theqtcompany.com>
Reviewed-by: Tim Jenssen <tim.jenssen@theqtcompany.com>