Add Ctrl+[n,p] Locator keyboard movement shortcuts.
On OS X, Ctrl+n and Ctrl+p, the Emacs bindings for next and previous,
are global throughout the system. In Qt Creator they work for menus,
code completion widget and other places like Vimium plugin, but not
the Locator widget. This change adds that functionality to Locator
too and makes navigating Locator items easier.

Change-Id: Iebce53d55841ba97d35ad2d433b0d9840d5e4137
Reviewed-by: David Schulz <david.schulz@theqtcompany.com>