Beautifier: Make all tools MIME restrict-able
The newly introduced auto save option is restricted to user definable
MIME types. The underlaying tool's restriction, however, is only if the
current editor is a cpp editor. This patch makes the tools also MIME
types restrict-able. In addition the auto save functionality is now only
applicable if the file matches the auto save MIME types as well as the
MIME types of the chosen tool.

Change-Id: Ic430b4a07341647e6c8e95d2b802a17db1637a36
Reviewed-by: Leena Miettinen <riitta-leena.miettinen@qt.io>