DocumentModel: Avoid deletion of entries for suspended documents
When opening a previously suspended document, reuse the existing entry
instead of replacing it. That way existing pointers to the entry are not
invalidated.

Fixes a regression introduced by 0c441563288ccbbffb8fcffa04f676398a6f2abb
that lead to crashes when selecting suspended documents from the
history popup (ctrl+tab / opt+tab) with the mouse.

Change-Id: Id41cae1132e4435c8e8bbd8624eaf840b02dc31f
Reviewed-by: Alexandru Croitor <alexandru.croitor@theqtcompany.com>
Reviewed-by: David Schulz <david.schulz@theqtcompany.com>